nil check

+++ app/models/cart.rb
@@ -24,11 +24,13 @@ class Cart < ActiveRecord::Base
 
   def remove_item(cartable, quantity = 1, owner=nil)
     existing_cart_item = cart_item_for_cartable_and_owner(cartable, owner)
-    Cart.transaction do
-      if existing_cart_item.quantity > quantity
-          existing_cart_item.update_attributes(:quantity => existing_cart_item.quantity - quantity)
-      else
-        existing_cart_item.destroy
+    if existing_cart_item
+      Cart.transaction do
+        if existing_cart_item.quantity > quantity
+            existing_cart_item.update_attributes(:quantity => existing_cart_item.quantity - quantity)
+        else
+          existing_cart_item.destroy
+        end
       end
     end
   end
